backend/requirements/development.txt
# =============================================================================
# STRATEKAZ - Dependencias de Desarrollo
# =============================================================================

# Base
-r base.txt

# -----------------------------------------------------------------------------
# Debug
# -----------------------------------------------------------------------------
django-debug-toolbar>=4.2,<4.3
django-extensions>=3.2,<3.3
ipython>=8.20,<8.21

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.0,<8.1
pytest-django>=4.7,<4.8
pytest-cov>=4.1,<4.2
pytest-asyncio>=0.23,<0.24
factory-boy>=3.3,<3.4
faker>=22.0,<23.0
responses>=0.24,<0.25
freezegun>=1.3,<1.4

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
flake8>=7.0,<7.1
black>=24.1,<24.2
isort>=5.13,<5.14
mypy>=1.8,<1.9
django-stubs>=4.2,<4.3
djangorestframework-stubs>=3.14,<3.15

# -----------------------------------------------------------------------------
# Pre-commit hooks
# -----------------------------------------------------------------------------
pre-commit>=3.6,<3.7

# -----------------------------------------------------------------------------
# Documentation
# -----------------------------------------------------------------------------
mkdocs>=1.5,<1.6
mkdocs-material>=9.5,<9.6

# -----------------------------------------------------------------------------
# Profiling
# -----------------------------------------------------------------------------
django-silk>=5.1,<5.2
py-spy>=0.3,<0.4
